Update eventtable to use util and move connectedevent out of folder
