+++ src/components/Router.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import EventSelector from './EventSelector'
-import ConnectedEvent from '../containers/ConnectedEvent'
+import ConnectedEvent from './ConnectedEvent'
 import Nav from './Nav'
 import { Auth } from "aws-amplify";
 import { Login, LoginRedirect } from './Authentication'
